+++ snug/__init__.py
@@ -1,4 +1,7 @@
 from .core import *  # noqa
+from .http import *  # noqa
+
+from . import http, clients  # noqa
 
 from .__about__ import (__author__, __copyright__,  # noqa
                         __version__, __description__)
