+++ tests/test_core.py
@@ -1,6 +1,5 @@
 import asyncio
 import inspect
-import json
 import sys
 import urllib.request
 from unittest import mock
@@ -11,11 +10,6 @@ import snug
 
 live = pytest.mark.skipif(not pytest.config.getoption('--live'),
                           reason='skip live data test')
-
-
-@pytest.fixture
-def loop(scope='module'):
-    return asyncio.get_event_loop()
 
 
 class MockAsyncClient:
@@ -254,79 +248,6 @@ class TestExecuteAsync:
             'my/url', headers={'Authorization': 'Bearer foo'})
 
 
-class TestRequest:
-
-    def test_defaults(self):
-        req = snug.Request('GET', 'my/url/')
-        assert req == snug.Request('GET', 'my/url/', params={}, headers={})
-
-    def test_with_headers(self):
-        req = snug.GET('my/url', headers={'foo': 'bla'})
-        assert req.with_headers({'other-header': 3}) == snug.GET(
-            'my/url', headers={'foo': 'bla', 'other-header': 3})
-
-    def test_with_prefix(self):
-        req = snug.GET('my/url/')
-        assert req.with_prefix('mysite.com/') == snug.GET(
-            'mysite.com/my/url/')
-
-    def test_with_params(self):
-        req = snug.GET('my/url/', params={'foo': 'bar'})
-        assert req.with_params({'other': 3}) == snug.GET(
-            'my/url/', params={'foo': 'bar', 'other': 3})
-
-    def test_equality(self):
-        req = snug.Request('GET', 'my/url')
-        other = req.replace()
-        assert req == other
-        assert not req != other
-
-        assert not req == req.replace(headers={'foo': 'bar'})
-        assert req != req.replace(headers={'foo': 'bar'})
-
-        assert not req == object()
-        assert req != object()
-
-    def test_repr(self):
-        req = snug.GET('my/url')
-        assert 'GET my/url' in repr(req)
-
-
-class TestResponse:
-
-    def test_equality(self):
-        rsp = snug.Response(204)
-        other = rsp.replace()
-        assert rsp == other
-        assert not rsp != other
-
-        assert not rsp == rsp.replace(headers={'foo': 'bar'})
-        assert rsp != rsp.replace(headers={'foo': 'bar'})
-
-        assert not rsp == object()
-        assert rsp != object()
-
-    def test_repr(self):
-        assert '404' in repr(snug.Response(404))
-
-
-def test_prefix_adder():
-    req = snug.GET('my/url')
-    adder = snug.prefix_adder('mysite.com/')
-    assert adder(req) == snug.GET('mysite.com/my/url')
-
-
-def test_header_adder():
-    req = snug.GET('my/url', headers={'Accept': 'application/json'})
-    adder = snug.header_adder({
-        'Authorization': 'my-auth'
-    })
-    assert adder(req) == snug.GET('my/url', headers={
-        'Accept': 'application/json',
-        'Authorization': 'my-auth'
-    })
-
-
 def test_executor():
     exec = snug.executor(client='foo')
     assert exec.keywords == {'client': 'foo'}
@@ -335,170 +256,6 @@ def test_executor():
 def test_async_executor():
     exec = snug.async_executor(client='foo')
     assert exec.keywords == {'client': 'foo'}
-
-
-def test_send_with_unknown_client():
-    class MyClass:
-        pass
-    with pytest.raises(TypeError, match='MyClass'):
-        snug.send(MyClass(), snug.GET('foo'))
-
-
-class TestSendWithUrllib:
-
-    @live
-    def test_no_contenttype(self):
-        req = snug.Request('POST', 'http://httpbin.org/post',
-                           content=b'foo',
-                           headers={'Accept': 'application/json'},
-                           params={'foo': 'bar'})
-        client = urllib.request.build_opener()
-        response = snug.send(client, req)
-        assert response == snug.Response(200, mock.ANY, headers=mock.ANY)
-        data = json.loads(response.content.decode())
-        assert data['args'] == {'foo': 'bar'}
-        assert data['data'] == 'foo'
-        assert data['headers']['Accept'] == 'application/json'
-        assert data['headers']['Content-Type'] == 'application/octet-stream'
-
-    @live
-    def test_no_data(self):
-        req = snug.Request('GET', 'http://httpbin.org/get',
-                           headers={'Accept': 'application/json'},
-                           params={'foo': 'bar'})
-        client = urllib.request.build_opener()
-        response = snug.send(client, req)
-        assert response == snug.Response(200, mock.ANY, headers=mock.ANY)
-        data = json.loads(response.content.decode())
-        assert data['args'] == {'foo': 'bar'}
-        assert data['headers']['Accept'] == 'application/json'
-        assert 'Content-Type' not in data['headers']
-
-    @live
-    def test_contenttype(self):
-        req = snug.Request('POST', 'http://httpbin.org/post',
-                           content=b'foo',
-                           headers={'content-Type': 'application/json'},
-                           params={'foo': 'bar'})
-        client = urllib.request.build_opener()
-        response = snug.send(client, req)
-        assert response == snug.Response(200, mock.ANY, headers=mock.ANY)
-        data = json.loads(response.content.decode())
-        assert data['args'] == {'foo': 'bar'}
-        assert data['data'] == 'foo'
-        assert data['headers']['Content-Type'] == 'application/json'
-
-
-def test_requests_send():
-    requests = pytest.importorskip("requests")
-    session = mock.Mock(spec=requests.Session)
-    req = snug.GET('https://www.api.github.com/organizations',
-                   params={'since': 3043},
-                   headers={'Accept': 'application/vnd.github.v3+json'})
-    req = snug.POST('http://httpbin.org/post',
-                    content=b'foo',
-                    params={'bla': 5},
-                    headers={'User-Agent': 'snug/dev'})
-    response = snug.send(session, req)
-    assert response == snug.Response(
-        status_code=session.request.return_value.status_code,
-        content=session.request.return_value.content,
-        headers=session.request.return_value.headers,
-    )
-    session.request.assert_called_once_with(
-        'POST',
-        'http://httpbin.org/post',
-        data=b'foo',
-        params={'bla': 5},
-        headers={'User-Agent': 'snug/dev'})
-
-
-def test_async_send_with_unknown_client(loop):
-    class MyClass:
-        pass
-
-    with pytest.raises(TypeError, match='MyClass'):
-        loop.run_until_complete(snug.send_async(MyClass(), snug.GET('foo')))
-
-
-@live
-class TestSendWithAsyncio:
-
-    def test_https(self, loop):
-        req = snug.Request('GET', 'https://httpbin.org/get',
-                           params={'param1': 'foo'},
-                           headers={'Accept': 'application/json'})
-        response = loop.run_until_complete(snug.send_async(loop, req))
-        assert response == snug.Response(200, mock.ANY, headers=mock.ANY)
-        data = json.loads(response.content.decode())
-        assert data['url'] == 'https://httpbin.org/get?param1=foo'
-        assert data['args'] == {'param1': 'foo'}
-        assert data['headers']['Accept'] == 'application/json'
-        assert data['headers']['User-Agent'].startswith('Python-asyncio/')
-
-    def test_http(self, loop):
-        req = snug.Request('POST', 'http://httpbin.org/post',
-                           content=json.dumps({"foo": 4}).encode(),
-                           headers={'User-agent': 'snug/dev'})
-        response = loop.run_until_complete(snug.send_async(loop, req))
-        assert response == snug.Response(200, mock.ANY, headers=mock.ANY)
-        data = json.loads(response.content.decode())
-        assert data['url'] == 'http://httpbin.org/post'
-        assert data['args'] == {}
-        assert json.loads(data['data']) == {'foo': 4}
-        assert data['headers']['User-Agent'] == 'snug/dev'
-
-
-@asyncio.coroutine
-def error(self):
-    yield from asyncio.sleep(0)
-    raise ValueError('foo')
-
-
-class TestAiohttpSend:
-
-    @live
-    def test_ok(self, loop):
-        req = snug.POST('https://httpbin.org/post',
-                        content=b'{"foo": 4}',
-                        params={'bla': 99},
-                        headers={'Accept': 'application/json'})
-        aiohttp = pytest.importorskip('aiohttp')
-
-        @asyncio.coroutine
-        def do_test():
-            session = aiohttp.ClientSession()
-            try:
-                return (yield from snug.send_async(session, req))
-            finally:
-                yield from session.close()
-
-        response = loop.run_until_complete(do_test())
-        assert response == snug.Response(200, mock.ANY, headers=mock.ANY)
-        data = json.loads(response.content.decode())
-        assert data['args'] == {'bla': '99'}
-        assert json.loads(data['data']) == {'foo': 4}
-        assert data['headers']['Accept'] == 'application/json'
-
-    @live
-    def test_error(self, loop):
-        req = snug.POST('https://httpbin.org/post',
-                        content=b'{"foo": 4}',
-                        params={'bla': 99},
-                        headers={'Accept': 'application/json'})
-        aiohttp = pytest.importorskip('aiohttp')
-
-        @asyncio.coroutine
-        def do_test():
-            session = aiohttp.ClientSession()
-            try:
-                return (yield from snug.send_async(session, req))
-            finally:
-                yield from session.close()
-
-        with mock.patch('aiohttp.client_reqrep.ClientResponse.read', error):
-            with pytest.raises(ValueError, match='foo'):
-                loop.run_until_complete(do_test())
 
 
 def test_relation():
@@ -531,144 +288,6 @@ def test_relation():
     assert isinstance(qux, f.Qux)
     qux2 = Foo.Qux(1, 2)
     assert isinstance(qux2, Foo.Qux)
-
-
-@pytest.fixture
-def headers():
-    return snug.Headers({
-        'Content-Type': 'application/json',
-        'accept': 'text/plain'
-    })
-
-
-class AlwaysEquals:
-    def __eq__(self, other):
-        return True
-
-    def __ne__(self, other):
-        return False
-
-
-class AlwaysInEquals:
-    def __eq__(self, other):
-        return False
-
-    def __ne__(self, other):
-        return True
-
-
-class TestHeaders:
-
-    def test_empty(self):
-        headers = snug.Headers()
-        assert len(headers) == 0
-        with pytest.raises(KeyError, match='foo'):
-            assert headers['foo']
-        assert list(headers) == []
-        assert headers == {}
-
-    def test_getitem(self, headers):
-        assert headers['Content-Type'] == 'application/json'
-        assert headers['content-type'] == 'application/json'
-        assert headers['Accept'] == 'text/plain'
-        assert headers['ACCEPT'] == 'text/plain'
-
-    def test_iter(self, headers):
-        assert set(headers) == {'Content-Type', 'accept'}
-
-    def test_equality(self, headers):
-        assert headers == headers
-        equivalent = {
-            'Accept': 'text/plain',
-            'Content-TYPE': 'application/json',
-        }
-        assert headers == equivalent
-        assert headers == snug.Headers(equivalent)
-        assert hash(headers) == hash(snug.Headers(equivalent))
-
-        assert not headers == {
-            'Content-Type': 'application/json',
-        }
-        assert not headers == snug.Headers({
-            'Content-Type': 'application/json',
-        })
-
-        assert not headers == {
-            'Content-TYPE': 'application/pdf',
-            'Accept': 'text/plain'
-        }
-        assert headers == AlwaysEquals()
-        assert not headers == AlwaysInEquals()
-
-    def test_inequality(self, headers):
-        assert not headers != headers
-        equivalent = {
-            'Accept': 'text/plain',
-            'Content-TYPE': 'application/json',
-        }
-        assert not headers != equivalent
-        assert not headers != snug.Headers(equivalent)
-        assert headers != {
-            'Content-Type': 'application/json',
-        }
-        assert headers != {
-            'Content-TYPE': 'application/pdf',
-            'Accept': 'text/plain'
-        }
-        assert hash(headers) != hash(snug.Headers({
-            'Content-TYPE': 'application/pdf',
-            'Accept': 'text/plain'
-        }))
-        assert not headers != AlwaysEquals()
-        assert headers != AlwaysInEquals()
-
-    def test_items(self, headers):
-        assert set(headers.items()) == {
-            ('Content-Type', 'application/json'),
-            ('accept', 'text/plain'),
-        }
-
-    def test_keys(self, headers):
-        assert 'Content-Type' in headers.keys()
-        assert 'Content-type' in headers.keys()
-        assert 'Content-Disposition' not in headers.keys()
-        assert set(headers.keys()) == {'Content-Type', 'accept'}
-
-    def test_len(self, headers):
-        assert len(headers) == 2
-
-    def test_bool(self, headers):
-        assert headers
-        assert not snug.Headers()
-
-    def test_contains(self, headers):
-        assert 'Content-Type' in headers
-        assert 'Content-type' in headers
-        assert 'ACCEPT' in headers
-        assert 'Content-Disposition' not in headers
-
-    def test_asdict(self, headers):
-        assert dict(headers) == {
-            'Content-Type': 'application/json',
-            'accept': 'text/plain'
-        }
-
-    def test_values(self, headers):
-        assert set(headers.values()) == {'application/json', 'text/plain'}
-
-    def test_get(self, headers):
-        assert headers.get('Content-Type', 'foo') == 'application/json'
-        assert headers.get('Content-TYPE', None) == 'application/json'
-        assert headers.get('Accept', None) == 'text/plain'
-        assert headers.get('Cookie', 'foo') == 'foo'
-
-    def test_repr(self, headers):
-        rep = repr(headers)
-        for key, value in headers.items():
-            assert key in rep
-            assert value in rep
-
-        assert repr(snug.Headers()) == '{<empty>}'
 
 
 def test_identity():
